+++ .gitignore
@@ -42,7 +42,7 @@ share/python-wheels/
 MANIFEST
 
 # PyInstaller
-#  Usually these files are written by a python script froms a template
+#  Usually these files are written by a python script forms a template
 #  before PyInstaller builds the exe, so as to inject date/other infos into it.
 *.manifest
 *.spec
