+++ apps/common/forms/multi_select.py
@@ -2,26 +2,27 @@
 """
     @project: maxkb
     @Author：虎
-    @file： table_radio.py
-    @date：2023/10/31 18:01
+    @file： multi_select.py
+    @date：2023/10/31 18:00
     @desc:
 """
-from typing import Dict
+from typing import List, Dict
 
-from common.froms.base_field import TriggerType, BaseExecField
+from common.forms.base_field import BaseExecField, TriggerType
 
 
-class TableRadio(BaseExecField):
+class MultiSelect(BaseExecField):
     """
-    table 单选
+    下拉单选
     """
 
     def __init__(self,
                  label: str,
                  text_field: str,
                  value_field: str,
-                 provider: str,
-                 method: str,
+                 option_list: List[str:object],
+                 provider: str = None,
+                 method: str = None,
                  required: bool = False,
                  default_value: object = None,
                  relation_show_field_dict: Dict = None,
@@ -29,5 +30,9 @@ class TableRadio(BaseExecField):
                  trigger_type: TriggerType = TriggerType.OPTION_LIST,
                  attrs: Dict[str, object] = None,
                  props_info: Dict[str, object] = None):
-        super().__init__("TableCheckbox", label, text_field, value_field, provider, method, required, default_value,
+        super().__init__("MultiSelect", label, text_field, value_field, provider, method, required, default_value,
                          relation_show_field_dict, relation_trigger_field_dict, trigger_type, attrs, props_info)
+        self.option_list = option_list
+
+    def to_dict(self):
+        return {**super().to_dict(), 'option_list': self.option_list}
