+++ apps/common/forms/password_input.py
@@ -2,16 +2,16 @@
 """
     @project: maxkb
     @Author：虎
-    @file： text_input_field.py
-    @date：2023/10/31 17:58
+    @file： password_input.py
+    @date：2023/11/1 14:48
     @desc:
 """
 from typing import Dict
 
-from common.froms.base_field import BaseField, TriggerType
+from common.forms import BaseField, TriggerType
 
 
-class TextInputField(BaseField):
+class PasswordInputField(BaseField):
     """
     文本输入框
     """
@@ -20,8 +20,7 @@ class TextInputField(BaseField):
                  required: bool = False,
                  default_value=None,
                  relation_show_field_dict: Dict = None,
-
                  attrs=None, props_info=None):
-        super().__init__('TextInput', label, required, default_value, relation_show_field_dict,
+        super().__init__('PasswordInput', label, required, default_value, relation_show_field_dict,
                          {},
                          TriggerType.OPTION_LIST, attrs, props_info)
