+++ apps/common/forms/radio_button_field.py
@@ -2,24 +2,25 @@
 """
     @project: maxkb
     @Author：虎
-    @file： array_object_card.py
-    @date：2023/10/31 18:03
+    @file： radio_field.py
+    @date：2023/10/31 17:59
     @desc:
 """
-from typing import Dict
+from typing import List, Dict
 
-from common.froms.base_field import BaseExecField, TriggerType
+from common.forms.base_field import BaseExecField, TriggerType
 
 
-class ArrayCard(BaseExecField):
+class Radio(BaseExecField):
     """
-    收集List[Object]
+    下拉单选
     """
 
     def __init__(self,
                  label: str,
                  text_field: str,
                  value_field: str,
+                 option_list: List[str:object],
                  provider: str,
                  method: str,
                  required: bool = False,
@@ -29,5 +30,9 @@ class ArrayCard(BaseExecField):
                  trigger_type: TriggerType = TriggerType.OPTION_LIST,
                  attrs: Dict[str, object] = None,
                  props_info: Dict[str, object] = None):
-        super().__init__("ArrayObjectCard", label, text_field, value_field, provider, method, required, default_value,
+        super().__init__("RadioButton", label, text_field, value_field, provider, method, required, default_value,
                          relation_show_field_dict, relation_trigger_field_dict, trigger_type, attrs, props_info)
+        self.option_list = option_list
+
+    def to_dict(self):
+        return {**super().to_dict(), 'option_list': self.option_list}
