+++ apps/common/forms/text_input_field.py
@@ -2,16 +2,16 @@
 """
     @project: maxkb
     @Author：虎
-    @file： password_input.py
-    @date：2023/11/1 14:48
+    @file： text_input_field.py
+    @date：2023/10/31 17:58
     @desc:
 """
 from typing import Dict
 
-from common.froms import BaseField, TriggerType
+from common.forms.base_field import BaseField, TriggerType
 
 
-class PasswordInputField(BaseField):
+class TextInputField(BaseField):
     """
     文本输入框
     """
@@ -20,7 +20,8 @@ class PasswordInputField(BaseField):
                  required: bool = False,
                  default_value=None,
                  relation_show_field_dict: Dict = None,
+
                  attrs=None, props_info=None):
-        super().__init__('PasswordInput', label, required, default_value, relation_show_field_dict,
+        super().__init__('TextInput', label, required, default_value, relation_show_field_dict,
                          {},
                          TriggerType.OPTION_LIST, attrs, props_info)
