+++ apps/setting/models_provider/impl/openai_model_provider/openai_model_provider.py
@@ -12,9 +12,9 @@ from typing import Dict
 from langchain.schema import HumanMessage
 from langchain_openai import ChatOpenAI
 
-from common import froms
+from common import forms
 from common.exception.app_exception import AppApiException
-from common.froms import BaseForm
+from common.forms import BaseForm
 from common.util.file_util import get_file_content
 from setting.models_provider.base_model_provider import IModelProvider, ModelProvideInfo, BaseModelCredential, \
     ModelInfo, \
@@ -50,8 +50,8 @@ class OpenAILLMModelCredential(BaseForm, BaseModelCredential):
     def encryption_dict(self, model: Dict[str, object]):
         return {**model, 'api_key': super().encryption(model.get('api_key', ''))}
 
-    api_base = froms.TextInputField('API 域名', required=True)
-    api_key = froms.PasswordInputField('API Key', required=True)
+    api_base = forms.TextInputField('API 域名', required=True)
+    api_key = forms.PasswordInputField('API Key', required=True)
 
 
 openai_llm_model_credential = OpenAILLMModelCredential()
